Move listener helper functions to separate file
